requirements.txt
faster_whisper
pyaudio
torchaudio
gpiozero